# Add Matrix.from_dod and simplify linsolve