--- a/sympy/polys/matrices/domainmatrix.py
+++ b/sympy/polys/matrices/domainmatrix.py
@@ -2538,7 +2538,6 @@ class DomainMatrix:
         See Also
         ========
 
-        particular
         rref
         rref_den
         solve_den_general