@oscarbenjamin 2023-12-19
> I want to make further changes to linsolve but it would start to depend on [gh-25996](https://github.com/sympy/sympy/pull/25996) which is not yet merged.

Actually this already depends on the changes there...