@oscarbenjamin 2023-12-20
So far this is showing some significant benchmark differences:
```
Significantly changed benchmark results (PR vs master)

| Change   | Before [4287ed72]    | After [46e287d1]    |   Ratio | Benchmark (Parameter)                               |
|----------|----------------------|---------------------|---------|-----------------------------------------------------|
| -        | 241±2ms              | 148±3ms             |    0.62 | baseline.TimeBaseline.time_loop10000000             |
| +        | 83.8±0.2μs           | 175±0.5μs           |    2.08 | solve.TimeRationalSystem.time_linsolve(1)           |
| +        | 217±0.9μs            | 337±2μs             |    1.55 | solve.TimeRationalSystem.time_linsolve(3)           |
| -        | 480±1μs              | 311±0.7μs           |    0.65 | solve.TimeSparseSystem.time_linear_eq_to_matrix(10) |
| -        | 900±2μs              | 568±2μs             |    0.63 | solve.TimeSparseSystem.time_linear_eq_to_matrix(20) |
| -        | 1.32±0.01ms          | 813±1μs             |    0.61 | solve.TimeSparseSystem.time_linear_eq_to_matrix(30) |
| -        | 10.5±0.1ms           | 4.22±0.01ms         |    0.4  | solve.TimeSparseSystem.time_linsolve_Aaug(10)       |
| -        | 27.9±0.2ms           | 14.8±0.04ms         |    0.53 | solve.TimeSparseSystem.time_linsolve_Aaug(20)       |
| -        | 53.1±0.2ms           | 31.3±0.05ms         |    0.59 | solve.TimeSparseSystem.time_linsolve_Aaug(30)       |
| -        | 10.3±0.09ms          | 4.29±0.03ms         |    0.42 | solve.TimeSparseSystem.time_linsolve_Ab(10)         |
| -        | 27.8±0.2ms           | 15.0±0.07ms         |    0.54 | solve.TimeSparseSystem.time_linsolve_Ab(20)         |
| -        | 52.7±0.2ms           | 31.9±0.2ms          |    0.61 | solve.TimeSparseSystem.time_linsolve_Ab(30)         |
| +        | 534±2μs              | 861±4μs             |    1.61 | solve.TimeSparseSystem.time_linsolve_eqs(10)        |
| +        | 979±3μs              | 1.53±0ms            |    1.57 | solve.TimeSparseSystem.time_linsolve_eqs(20)        |
| +        | 1.42±0.01ms          | 2.22±0.01ms         |    1.56 | solve.TimeSparseSystem.time_linsolve_eqs(30)        |
| -        | 11.9±0.2ms           | 7.89±0.07ms         |    0.66 | solve.TimeSparseSystem.time_solve(10)               |
```
This is because of a number of different factors:

Some solve functions used solve_lin_sys which uses dense DomainMatrix RREF and others used _linsolve which uses sparse RREF. The change here means that all solvers will use sparse RREF which makes many cases faster. The solve_lin_sys solver uses the connected components decomposition for the system which probably makes some cases faster though so potentially switching to the sparse solver does slow some cases down a bit even if it makes others faster.

The slowdowns seen in some cases in the benchmarks are for linsolve when passed equations as input. This needs further investigation but is probably not hard to improve. It might be related to using `linear_eq_to_matrix`...